cdc-agent/requirements.txt
# CDC Agent dependencies
confluent-kafka==2.6.1
hdbcli==2.21.28
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
prometheus-client==0.21.1
redis==5.2.1
